:recycle: Reduce code duplication
