+++ src/components/FormStart/index.js
@@ -1,25 +1,10 @@
-import React, {useEffect} from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import {FormattedMessage, useIntl} from 'react-intl';
+import {useIntl} from 'react-intl';
 
-import AuthenticationOutage, { useDetectAuthenticationOutage } from 'components/auth/AuthenticationOutage';
-import {useDetectAuthErrorMessages, AuthenticationErrors} from 'components/auth/AuthenticationErrors';
-import Body from 'components/Body';
-import Button from 'components/Button';
-import Card from 'components/Card';
-import {Literal, LiteralsProvider} from 'components/Literal';
-import MaintenanceMode from 'components/MaintenanceMode';
-import { Toolbar, ToolbarList } from 'components/Toolbar';
-import useQuery from 'hooks/useQuery';
 import Types from 'types';
-import LoginButton, {START_FORM_QUERY_PARAM} from 'components/LoginButton';
-import LoginButtonIcons from 'components/LoginButtonIcons';
-
-
-const useStartSubmission = () => {
-  const query = useQuery();
-  return !!query.get(START_FORM_QUERY_PARAM);
-};
+import {START_FORM_QUERY_PARAM} from 'components/LoginButton';
+import FormLanding from 'components/FormLanding';
 
 
 export const getLoginUrl = (loginOption) => {
@@ -39,35 +24,10 @@ export const getLoginUrl = (loginOption) => {
  * Form start screen.
  *
  * This is shown when the form is initially loaded and provides the explicit user
- * action to start the form, or (in the future) present the login button (DigiD,
- * eHerkenning...)
+ * action to start the form, or present the login button (DigiD, eHerkenning, eIDAS)
  */
 const FormStart = ({ form, onFormStart }) => {
   const intl = useIntl();
-  const doStart = useStartSubmission();
-  const outagePluginId = useDetectAuthenticationOutage();
-  const authErrors = useDetectAuthErrorMessages();
-  const hasAuthErrors = !!outagePluginId || !!authErrors;
-
-  useEffect(() => {
-    if (doStart && !hasAuthErrors) onFormStart();
-  }, [doStart, outagePluginId, hasAuthErrors, onFormStart]);
-
-  if (form.maintenanceMode) {
-    return <MaintenanceMode title={form.name} />;
-  }
-
-  if (outagePluginId) {
-    const loginOption = form.loginOptions.find(option => option.identifier === outagePluginId);
-    if (!loginOption) throw new Error('Unknown login plugin identifier');
-    return (
-      <Card title={
-        <FormattedMessage description="Form start outage title" defaultMessage="Problem - {formName}" values={{formName: form.name}} />
-      }>
-        <AuthenticationOutage loginOption={loginOption} />
-      </Card>
-    );
-  }
 
   const canLogin = form.loginOptions.length > 0;
   const startLoginMessage = form.loginRequired
@@ -85,44 +45,14 @@ const FormStart = ({ form, onFormStart }) => {
          defaultMessage: 'Please click the button below to start the form.'
       })
   ;
+
   return (
-    <LiteralsProvider literals={form.literals}>
-      <Card title={form.name}>
-
-        { !!authErrors ? <AuthenticationErrors parameters={authErrors}/> : null }
-
-        <Body modifiers={['compact']}>
-          {startLoginMessage}
-        </Body>
-
-        <Toolbar modifiers={['start']}>
-          <ToolbarList>
-            { form.loginRequired
-              ? null
-              : (
-                  <Button variant="primary" component="a" href="#" onClick={onFormStart}>
-                    <Literal name="beginText" />
-                  </Button>
-              )
-            }
-            {
-              form.loginOptions.map((option) => (
-                <LoginButton
-                  key={option.identifier}
-                  option={option}
-                  getLoginUrl={getLoginUrl}
-                />
-              ))
-            }
-          </ToolbarList>
-
-          <ToolbarList>
-            <LoginButtonIcons loginOptions={form.loginOptions} />
-          </ToolbarList>
-        </Toolbar>
-
-      </Card>
-    </LiteralsProvider>
+    <FormLanding
+      form={form}
+      onFormStart={onFormStart}
+      getLoginUrl={getLoginUrl}
+      startLoginMessage={startLoginMessage}
+    />
   );
 };
